+++ lib/inttegro/payouts.ex
@@ -321,4 +321,23 @@ defmodule Inttegro.Payouts do
       {:ok, Inttegro.Payouts.Payout.from_map(value)}
     end
   end
+
+  @doc "Search payouts with a typed resource-local query."
+  @spec search(Client.t(), Inttegro.Search.Request.t(), keyword()) ::
+          {:ok, Inttegro.Search.Page.t()} | {:error, Exception.t()}
+  def search(client, request, options \\ []) do
+    with {:ok, value} <-
+           Client.request(
+             client,
+             "POST",
+             "/payouts/search",
+             Inttegro.Codec.encode(request),
+             options,
+             "payouts.search",
+             "search",
+             true
+           ) do
+      {:ok, Inttegro.Search.Page.from_map(value)}
+    end
+  end
 end