+++ lib/inttegro/products.ex
@@ -354,4 +354,23 @@ defmodule Inttegro.Products do
       {:ok, Inttegro.Products.Page.from_map(value)}
     end
   end
+
+  @doc "Search products with a typed resource-local query."
+  @spec search(Client.t(), Inttegro.Search.Request.t(), keyword()) ::
+          {:ok, Inttegro.Search.Page.t()} | {:error, Exception.t()}
+  def search(client, request, options \\ []) do
+    with {:ok, value} <-
+           Client.request(
+             client,
+             "POST",
+             "/products/search",
+             Inttegro.Codec.encode(request),
+             options,
+             "products.search",
+             "search",
+             true
+           ) do
+      {:ok, Inttegro.Search.Page.from_map(value)}
+    end
+  end
 end